Move RealityDataSource into its own file. Add a RealityDataConnectionManager that will cached created connection so reality data doesn't need to be fetch every time a reality data is detached and then re-attached.
